Remove unused modules (#590)
* Remove unused modules'

* rename more modules, remove public module

* Restore serde_structs

* use local starknet-rs in cli
