+++ src/core/mod.rs
@@ -1,4 +1,3 @@
-pub mod block_hash;
 pub mod contract_address;
 pub mod errors;
 pub mod syscalls;
