+++ src/storage/errors/storage_errors.rs
