+++ tests/internals.rs
@@ -11,7 +11,9 @@ use lazy_static::lazy_static;
 use num_traits::{Num, One, ToPrimitive, Zero};
 use starknet_contract_class::EntryPointType;
 use starknet_rs::core::errors::state_errors::StateError;
-use starknet_rs::definitions::constants::DEFAULT_CAIRO_RESOURCE_FEE_WEIGHTS;
+use starknet_rs::definitions::constants::{
+    DEFAULT_CAIRO_RESOURCE_FEE_WEIGHTS, VALIDATE_ENTRY_POINT_SELECTOR,
+};
 use starknet_rs::services::api::contract_classes::deprecated_contract_class::ContractClass;
 use starknet_rs::{
     business_logic::{
@@ -44,7 +46,6 @@ use starknet_rs::{
         general_config::{StarknetChainId, StarknetGeneralConfig, StarknetOsConfig},
         transaction_type::TransactionType,
     },
-    public::abi::VALIDATE_ENTRY_POINT_SELECTOR,
     utils::{calculate_sn_keccak, felt_to_hash, Address, ClassHash},
 };
 use std::{
